Demographics changes:
(3) Demographics.
(i) Enable a user to electronically record, change, and access patient
demographic data including preferred language, sex, race, ethnicity, and
date of birth.
(A) Enable race and ethnicity to be recorded in accordance with the
standard specified in § 170.207(f) and whether a patient declines to
specify race and/or ethnicity.
(B) Enable preferred language to be recorded in accordance with the
standard specified in § 170.207(g) and whether a patient declines to
specify a preferred language. - See more at:
http://www.open-emr.org/wiki/index.php/Demographics_(MU2)#sthash.uTvBfKjH.dpuf

Demographics changes - Post Review
Changes:
1. Add new column (list_backup_id) to layout_options table
2. get backup list from frow
3. Add column back up list in layouts editor
4. Change column name to ISO 639-2 code instead of notes for languages
5. no upgrade script changes in this check in (upgrade script will be on
next check in)

database script changes - post review